Custom keymap-drawer CI workflow
Rather than using the stock CI workflow for keymap-drawer, implement a
custom workflow that supports generating a seperate SVG for each layer.

Untested and incomplete. The goal is to have a seperate job for drawing
each layer, store the resulting SVGs as artifacts, and then finally
commit them all in one go.